requirements.txt
anthropic>=0.6.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
jsonschema>=4.17.3
Flask==2.2.3
Werkzeug==2.2.3
